Initial gnome_shell support
Gnome shell support seems to be working ...in fact it seems to be much better than on Unity. Since Gnome-shell doesn't use viewports but desktops, the user doesn't have to jump around from screen to screen to place windows. Also seems a bit snappier.
refactored both wm.py, wmctrl.py, workset.py and unitywm.py a bit to account for the differences, but more can be done to clean up.